+++ client/packages/react-native/src/Storage.js
@@ -1,15 +1,3 @@
-import AsyncStorage from "@react-native-async-storage/async-storage";
+import { IndexedDBStorage } from "@instantdb/core";
 
-export default class Storage {
-  constructor(dbName) {
-    this.dbName = dbName;
-  }
-
-  async getItem(k) {
-    return await AsyncStorage.getItem(`${this.dbName}_${k}`);
-  }
-
-  async setItem(k, v) {
-    await AsyncStorage.setItem(`${this.dbName}_${k}`, v);
-  }
-}
+export default IndexedDBStorage;
