--- client/packages/react-native/src/Storage.web.js
@@ -1,3 +1,0 @@
-import { IndexedDBStorage } from "@instantdb/core";
-
-export default IndexedDBStorage;
